+++ client-course-schedulizer/src/components/ColorSelector/index.ts
@@ -1,0 +1,1 @@
+export * from "./ColorSelector";
