+++ client-course-schedulizer/src/components/ScheduleToolbar/ScheduleToolbar.scss
@@ -2,9 +2,3 @@
   display: flex;
   justify-content: space-between;
 }
-
-.semester-selector {
-  display: flex;
-  text-align: center;
-  align-items: center;
-}
