+++ client-course-schedulizer/src/components/Searchbar/index.ts
@@ -1,0 +1,1 @@
+export * from "./Searchbar";
