+++ client-course-schedulizer/src/components/SemesterSelector/SemesterSelector.scss
@@ -1,0 +1,5 @@
+.semester-selector {
+  display: flex;
+  text-align: center;
+  align-items: center;
+}
