+++ client-course-schedulizer/src/components/SemesterSelector/index.ts
@@ -1,0 +1,1 @@
+export * from "./SemesterSelector";
